requirements.txt
flask==2.0.1
flask-cors==3.0.10
supabase==1.0.3
python-dotenv==0.19.0
authlib==1.2.0
gunicorn==20.1.0
